+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API/UseCases/Dashboard/GetDashboardService.cs
@@ -1,4 +1,4 @@
-﻿using Dfe.ManageFreeSchoolProjects.API.Contracts.Dashboard;
+using Dfe.ManageFreeSchoolProjects.API.Contracts.Dashboard;
 using Dfe.ManageFreeSchoolProjects.API.Contracts.Project;
 using Dfe.ManageFreeSchoolProjects.API.Extensions;
 using Dfe.ManageFreeSchoolProjects.API.UseCases.Project;
@@ -35,10 +35,9 @@ namespace Dfe.ManageFreeSchoolProjects.API.UseCases.Dashboard
 
             query = ApplyFilters(query, parameters);
 
-            var count = query.Count();
+            var count = await query.CountAsync();
             
-            var projectRecords = await 
-                query
+            var projectRecords = await query
                     .OrderByDescending(kpi => kpi.ProjectStatusProvisionalOpeningDateAgreedWithTrust)
                     .ThenBy(kpi => kpi.ProjectStatusCurrentFreeSchoolName)
                     .Paginate(parameters.Page, parameters.Count)
@@ -108,14 +107,12 @@ namespace Dfe.ManageFreeSchoolProjects.API.UseCases.Dashboard
 
             query = ApplyFilters(query, parameters);
 
-            await query
-                    .OrderByDescending(kpi => kpi.ProjectStatusProvisionalOpeningDateAgreedWithTrust)
-                    .ThenBy(kpi => kpi.ProjectStatusCurrentFreeSchoolName)
-                    .ToListAsync();
+            var totalListOfIds = await query
+                .Select(x => x.ProjectStatusProjectId)
+                .Distinct()
+                .ToListAsync();
 
-            var totalListOfIds = query.Select(x => x.ProjectStatusProjectId).Distinct();
-
-            return await totalListOfIds.ToListAsync();
+            return totalListOfIds;
         }
     }
 }